Hybrid_Search/requirements.txt
# Core
python-dotenv>=1.0.1
numpy>=1.26.0

# Pinecone (new SDK)
pinecone>=5.0.0
pinecone-text>=0.9.0
pinecone-notebooks

# LangChain + HF embeddings
langchain>=0.2.12
langchain-community>=0.2.10
langchain-huggingface>=0.0.3
# (langchain-huggingface pulls sentence-transformers)

# Required for HuggingFaceEmbeddings
sentence-transformers>=2.2.2